requirements.txt
google-cloud-translate==2.0.1
keyboard==0.13.5
pyperclip==1.9.0
pyautogui==0.9.54